app/requirements.txt
fastapi == 0.120.0
uvicorn[standard] == 0.38.0
python-dotenv == 1.1.1
pydantic == 2.12.3
httpx == 0.28.1
openai == 2.6.1
google-api-python-client == 2.149.0
google-auth-httplib2 == 0.2.0
google-auth-oauthlib == 1.2.1
pytest == 8.3.3
pytest-asyncio == 0.24.0